requirements.txt
openai==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
supabase==1.0.3
stripe==6.0.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.3.0
